发布最新正式版4.2.0

+++ weixin-java-cp/src/main/java/me/chanjar/weixin/cp/bean/external/WxCpUserExternalTagGroupInfo.java
@@ -35,7 +35,7 @@ public class WxCpUserExternalTagGroupInfo extends WxCpBaseResp {
     private Long createTime;
 
     @SerializedName("order")
-    private Integer order;
+    private Long order;
 
     @SerializedName("deleted")
     private Boolean deleted;
@@ -64,13 +64,13 @@ public class WxCpUserExternalTagGroupInfo extends WxCpBaseResp {
     private String name;
 
     @SerializedName("create_time")
-    private Long  createTime;
+    private Long createTime;
 
     @SerializedName("order")
-    private Integer  order;
+    private Long order;
 
     @SerializedName("deleted")
-    private Boolean  deleted;
+    private Boolean deleted;
 
   }
